Separate NextPrevLink for reuse
