+++ src/components/InternalPageNav/InternalPageNav.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Button from '@leafygreen-ui/button';
-import { Body } from '@leafygreen-ui/typography';
-import Icon, { glyphs } from '@leafygreen-ui/icon';
+import { glyphs } from '@leafygreen-ui/icon';
 import { css, cx } from '@leafygreen-ui/emotion';
-import { palette } from '@leafygreen-ui/palette';
 import { theme } from '../../theme/docsTheme';
 import { getPageTitle } from '../../utils/get-page-title';
 import Link from '../Link';
+import NextPrevLink from './NextPrevLink';
 
 const containerStyling = css`
   padding-bottom: 2.5em;
@@ -27,62 +25,6 @@ const containerStyling = css`
     display: none;
   }
 `;
-
-const commonTextStyles = css`
-  font-size: ${theme.fontSize.small};
-  line-height: 20px;
-`;
-
-const nextPrevTextStyling = css`
-  ${commonTextStyles}
-  font-weight: 500;
-`;
-
-const nextTextStyling = css`
-  text-align: end;
-`;
-
-const prevTextStyling = css`
-  text-align: start;
-`;
-
-const nextPrevTitleTextStyling = css`
-  ${commonTextStyles}
-  color: ${palette.gray.base};
-`;
-
-const commonLinkContentContainerStyling = css`
-  align-items: center;
-  display: flex;
-  column-gap: 14px;
-`;
-
-const nextLinkContainerStyling = css`
-  ${commonLinkContentContainerStyling}
-  flex-direction: row-reverse;
-`;
-
-const prevLinkContainerStyling = css`
-  ${commonLinkContentContainerStyling}
-  flex-direction: row;
-`;
-
-const NextPrevLink = ({ icon, direction, pageTitle }) => {
-  const isNext = direction?.toLowerCase() === 'next';
-  const isPrev = direction?.toLowerCase() === 'back';
-
-  return (
-    <div className={cx({ [nextLinkContainerStyling]: isNext, [prevLinkContainerStyling]: isPrev })}>
-      <Button size="large">
-        <Icon glyph={icon} />
-      </Button>
-      <div className={cx({ [nextTextStyling]: isNext }, { [prevTextStyling]: isPrev })}>
-        <Body className={cx(nextPrevTextStyling)}>{direction}</Body>
-        <Body className={cx(nextPrevTitleTextStyling)}>{pageTitle}</Body>
-      </div>
-    </div>
-  );
-};
 
 const InternalPageNav = ({ slug, slugTitleMapping, toctreeOrder }) => {
   const slugIndex = toctreeOrder.indexOf(slug);
